refactor: Renames components, and merges header name in the header component.
