+++ src/components/header.tsx
@@ -3,15 +3,15 @@
 import { useSideBarToggle } from '@/hooks/useSidebarToggle';
 import { usePathname } from 'next/navigation';
 import React from 'react';
-import HeaderName from './headerName';
-import HeaderOps from './headerOps';
+import HeaderOps from './HeaderOps';
 import Icon from "../lib/constants/icons";
 import Image from 'next/image';
-import HeaderSearch from './headerSearch';
+import HeaderSearch from './HeaderSearch';
+import { routes } from '@/lib/constants';
 
 const Header = () => {
   const { isSidebarVisible } = useSideBarToggle();
-  const currentPath = usePathname();
+  const currentRoute = usePathname();
 
   return (
     <div className={`bg-white sticky top-0 grid grid-cols-1 place-items-center h-[21vh] md:h-[8vh] w-full z-30 md:pr-8 transition-all duration-300 ease-in-out pl-5 md:pl-[13rem]`}>
@@ -21,15 +21,23 @@ const Header = () => {
         </div>
 
         <div className="col-span-1 md:col-span-2 flex justify-end md:justify-normal items-center pr-6 md:pr-0">
-          <HeaderName />
+        <div>
+        {routes.sidebarItems.map((item, index) => (
+              <div key={index}>
+                  <h2 className="text-3xl sm:text3xl font-bold">
+                    {currentRoute === item.path && <div className=''>{item.title}</div>}
+                  </h2>
+              </div>
+            ))}
+    </div>
         </div>
 
         <div className="col-span-1 md:col-span-3 flex justify-start md:justify-normal items-center">
-          <HeaderSearch currentPath={currentPath} />
+          <HeaderSearch currentPath={currentRoute} />
         </div>
 
         <div className="col-span-1 md:col-span-3 flex justify-end md: items-center">
-          <HeaderOps currentPath={currentPath} />
+          <HeaderOps currentPath={currentRoute} />
         </div>
       </div>
     </div>
